backend/foodgram_backend/requirements.txt
Django==3.2.15
djangorestframework==3.13.1
psycopg2-binary==2.8.6
Pillow==9.2.0
django-extra-fields==0.3
djoser
djangorestframework-simplejwt==4.7.2